refactor: BindingAdapter 시간 변환을 TimeConverter에서 진행
